requirements.txt
streamlit
langchain
langchain-google-genai
pypdf
python-dotenv
